+++ src/file_pipe_log/pipe_builder.rs
@@ -16,11 +16,11 @@ use crate::config::{Config, RecoveryMode};
 use crate::env::FileSystem;
 use crate::event_listener::EventListener;
 use crate::log_batch::LogItemBatch;
-use crate::pipe_log::{FileId, FileSeq, LogQueue};
+use crate::pipe_log::{FileId, FileSeq, LogFileContext, LogQueue, Version};
 use crate::util::Factory;
 use crate::{Error, Result};
 
-use super::format::{lock_file_path, FileNameExt, LogFileContext, LogFileFormat, Version};
+use super::format::{lock_file_path, FileNameExt, LogFileFormat};
 use super::log_file::{build_file_reader, FileHandler};
 use super::pipe::{DualPipes, SinglePipe};
 use super::reader::LogItemBatchFileReader;
@@ -420,8 +420,6 @@ impl<F: FileSystem> DualPipesBuilder<F> {
             files,
             match queue {
                 LogQueue::Append => self.cfg.recycle_capacity(),
-                // No need to build RecycleFileCollection for
-                // LogQueue::Rewrite.
                 LogQueue::Rewrite => 0,
             },
         )